requirements.txt
torch>=2.0.0
transformers>=4.40.0
qwen-vl-utils>=0.0.1
Pillow>=10.0.0
requests>=2.31.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
tiktoken>=0.5.0